chore: move formatUsdValue to receipt feature
