+++ app/utils/index.ts
@@ -61,15 +61,6 @@ export function lamportsToSolString(lamports: number | bigint, maximumFractionDi
     return new Intl.NumberFormat('en-US', { maximumFractionDigits }).format(sol);
 }
 
-export function formatUsdValue(amount: number, price: number): string {
-    const value = amount * price;
-    if (isNaN(value) || value < 0) return '~0.00 USD';
-    return `~${value.toLocaleString('en-US', {
-        maximumFractionDigits: 2,
-        minimumFractionDigits: 2,
-    })} USD`;
-}
-
 export function numberWithSeparator(s: string) {
     // eslint-disable-next-line no-restricted-syntax -- insert thousands separator
     return s.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
